Remove fs persistance
Currently the app was persisting user info in a json file. After
deploying the app on heroku it was found that when the app changes
its state to idle, heroku was resetting the app files and thus
losing the persisted data.

At the moment persistance has been removed, but in the future we
may use a mongodb addon. ^_^